Update resume to have a two-column format
I realized that a lot of space was being wasted with the colorboxes for
resume item headers, which could easily be recovered by using two columns.
Experimented a bit with breaking it apart into sections, some in one column and
some in two, but found that doing two column throughout was the most effective
approach. Required some re-jiggering of various bits of content and formatting
to get the look right.