Add JSON version of resume content
The JSON document includes a lot of tex-related metadata, but as much as
possible that is encoded as "style" information about sections. Those styles
could be used as CSS classes or similar in another target, which allows that
information to be rendered correctly in latex without being overly specific.
There is still one unfortunate case of "tex" specific metadata -- namely
handling of elements which should be wrapped in \url{}. Another solution is
needed there -- likely converting every element to a JSON object with "type:
text" implicitly included.